requirements.txt
                    # Reddit API
praw>=7.7.1

# LLM Providers
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.4.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
pydantic>=2.5.0
rich>=13.7.0
pyyaml>=6.0.0

# Logging and monitoring
structlog>=23.2.0

# Web UI
flask>=3.0.0
flask-cors>=4.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0